// Features/reports/Types/IPunchForReport.ts
export interface IPunchForReport {
    identifier: string;
    idEmployee: number;
    employeeName: string;
    workScheduler: string;
    idPosition: number;
    idDepartment: number;
    position: string;
    department: string;
    punchDate: string;
    hourIn: string;
    hourOut: string;
    isManual: string;
    idCompany: number;
}